% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3rain.R
\name{chart_settings}
\alias{chart_settings}
\title{Adjust chart settings}
\usage{
chart_settings(d3rain, toolTipTextColor = "white",
  backgroundFill = "white", fontSize = 18, fontFamily = "sans-serif",
  yAxisTickLocation = "center")
}
\arguments{
\item{d3rain}{An object of class d3rain}

\item{toolTipTextColor}{Color of tooltip text}

\item{backgroundFill}{Background color of SVG}

\item{fontSize}{Font size}

\item{fontFamily}{Font family, e.g. 'times', 'sans-serif'}

\item{yAxisTickLocation}{Location of y-axis ticks, either 'center', 'left', or 'right'}
}
\value{
d3rain
}
\description{
Adjust chart settings
}
\examples{
mtcars$cyl <- factor(mtcars$cyl)
mtcars$car <- rownames(mtcars)
d3rain(mtcars, mpg, cyl, car) \%>\%
    drip_settings(ease = 'linear', jitterWidth = 25, dripSpeed = 500) \%>\%
    chart_settings(fontFamily = 'times', yAxisTickLocation = 'left')
}
