% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3rain.R
\name{d3rain}
\alias{d3rain}
\title{Create a d3rain visualization}
\usage{
d3rain(.data, x, y, toolTip, reverseX = FALSE, title = "")
}
\arguments{
\item{.data}{A table of data.}

\item{x}{A numeric 'ranking' variable, e.g. percentile, rank, etc.}

\item{y}{A factored, ordinal variable.}

\item{toolTip}{Which variable to display on drip tooltips.}

\item{title}{Visualization title}
}
\description{
Create a d3rain visualization
}
