% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3rain_hist.R
\name{d3rain_hist}
\alias{d3rain_hist}
\title{Create Rain Histograms}
\usage{
d3rain_hist(.data, x, levels, title = "", xBins = 40)
}
\arguments{
\item{.data}{a table of data}

\item{x}{the binning x variable}

\item{levels}{the ordered columns}

\item{title}{chart title}

\item{xBins}{Number of bins. Defaults to 40.}
}
\value{
an object of class 'd3rain_hist'
}
\description{
Create Rain Histograms
}
\examples{
data.frame(
x = rnorm(100),
l1 = sample(c(TRUE, FALSE), replace = TRUE, size = 100, prob = c(0.8, 0.2)),
l2 = sample(c(TRUE, FALSE), replace = TRUE, size = 100, prob = c(0.5, 0.5)),
l3 = sample(c(TRUE, FALSE), replace = TRUE, size = 100, prob = c(0.3, 0.7))
) \%>\%
 d3rain_hist(x = x, levels = c("l1", "l2", "l3"), title = "my title")
}
