% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3rain.R
\name{drip_behavior}
\alias{drip_behavior}
\title{Adjust drip behavior}
\usage{
drip_behavior(d3rain, dripSequence = "iterate", ease = "bounce",
  dripSpeed = 1500, iterationSpeedX = 100, jitterWidth = 0)
}
\arguments{
\item{d3rain}{An object of class d3rain}

\item{dripSequence}{Either 'iterate' or 'together'}

\item{ease}{Either 'bounce' or 'linear'}

\item{dripSpeed}{Drip speed}

\item{iterationSpeedX}{Iteration speed multiplier}

\item{jitterWidth}{Jitter width in pixels along x-axis}
}
\description{
Adjust drip behavior
}
\examples{
mtcars$cyl <- factor(mtcars$cyl)
mtcars$car <- rownames(mtcars)
d3rain(mtcars, mpg, cyl, car) \%>\%
    drip_behavior(ease = 'linear', jitterWidth = 25, dripSpeed = 500)
}
