% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3rain.R
\name{drip_settings}
\alias{drip_settings}
\title{Adjust drip style and behavior}
\usage{
drip_settings(d3rain, dripFill = "firebrick", dripOpacity = 0.5,
  dripSequence = "iterate", ease = "bounce", dripSpeed = 1500,
  iterationSpeedX = 100, jitterWidth = 0)
}
\arguments{
\item{d3rain}{An object of class d3rain}

\item{dripFill}{Color of drips}

\item{dripOpacity}{Opacity of drips}

\item{dripSequence}{Either 'iterate' or 'together'}

\item{ease}{Either 'bounce' or 'linear'}

\item{dripSpeed}{Drip speed in milliseconds}

\item{iterationSpeedX}{Iteration speed multiplier}

\item{jitterWidth}{Jitter width in pixels along x-axis}
}
\description{
Adjust drip style and behavior
}
\examples{
mtcars$cyl <- factor(mtcars$cyl)
mtcars$car <- rownames(mtcars)
d3rain(mtcars, mpg, cyl, car) \%>\%
    drip_settings(ease = 'linear', jitterWidth = 25, dripSpeed = 500)
}
