% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3rain.R
\name{drip_style}
\alias{drip_style}
\title{Adjust drip style}
\usage{
drip_style(d3rain, dripFill = "firebrick", toolTipTextColor = "white",
  backgroundFill = "white", fontSize = 18, fontFamily = "sans-serif",
  dripOpacity = 0.5, yAxisTickLocation = "center")
}
\arguments{
\item{d3rain}{An object of class d3rain}

\item{dripFill}{Color of drips}

\item{backgroundFill}{Background color of SVG}

\item{fontSize}{Font size}

\item{fontFamily}{Font family, e.g. 'times', 'sans-serif'}

\item{dripOpacity}{Opacity of drips}

\item{yAxisTickLocation}{Location of y-axis ticks, either 'center', 'left', or 'right'}
}
\value{
d3rain
}
\description{
Adjust drip style
}
\examples{
mtcars$cyl <- factor(mtcars$cyl)
mtcars$car <- rownames(mtcars)
d3rain(mtcars, mpg, cyl, car) \%>\%
    drip_behavior(ease = 'linear', jitterWidth = 25, dripSpeed = 500) \%>\%
    drip_style(dripFill = 'steelblue', dripOpacity = 0.2)
}
