% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3rain.R
\name{drop_behavior}
\alias{drop_behavior}
\title{Adjust d3rain drop behavior}
\usage{
drop_behavior(d3rain, dropSequence = "iterate", ease = "bounce",
  dropSpeed = 1500, iterationSpeedX = 100, jitterWidth = 0)
}
\arguments{
\item{d3rain}{An object of class d3rain}

\item{ease}{Either 'bounce' or 'linear'}

\item{dropSpeed}{Drop speed}

\item{iterationSpeedX}{Iteration speed multiplier}
}
\description{
Adjust d3rain drop behavior
}
