% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3rain.R
\name{drop_style}
\alias{drop_style}
\title{Adjust d3rain drop style}
\usage{
drop_style(d3rain, dropFill = "firebrick", backgroundFill = "white",
  fontSize = 18, fontFamily = "sans-serif", dropOpacity = 0.5)
}
\arguments{
\item{d3rain}{An object of class d3rain}

\item{dropFill}{Color of drops}

\item{backgroundFill}{Background color of SVG}

\item{fontSize}{Font size}

\item{fontFamily}{Font family, e.g. 'times', 'sans-serif'}
}
\description{
Adjust d3rain drop style
}
