% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3rain_hist.R
\name{hist_drip_settings}
\alias{hist_drip_settings}
\title{Adjust hist drip settings}
\usage{
hist_drip_settings(d3rain_hist, colors = NULL,
  transitionIntervals = 2500, dripSpeed = 300, dripSize = 5)
}
\arguments{
\item{d3rain_hist}{an object of d3rain_hist}

\item{colors}{a vector of colors}

\item{transitionIntervals}{milliseconds between group transitions}

\item{dripSpeed}{drop speed}

\item{dripSize}{drip radius}
}
\value{
an object of d3rain_hist
}
\description{
Adjust hist drip settings
}
\examples{
data.frame(
x = rnorm(100),
l1 = sample(c(TRUE, FALSE), replace = TRUE, size = 100, prob = c(0.8, 0.2)),
l2 = sample(c(TRUE, FALSE), replace = TRUE, size = 100, prob = c(0.5, 0.5)),
l3 = sample(c(TRUE, FALSE), replace = TRUE, size = 100, prob = c(0.3, 0.7))
) \%>\%
 d3rain_hist(x = x, levels = c("l1", "l2", "l3"), title = "my title") \%>\%
 hist_chart_settings(annotations = c("Annotation1", "Annotation2", "YO!"),
                     titlePosition = "left",
                     levelLabelLocation = "left") \%>\%
 hist_drip_settings(colors = c("red", "blue", "green"))
}
