% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3rain.R
\name{drip_behavior}
\alias{drip_behavior}
\title{Adjust drip behavior}
\usage{
drip_behavior(d3rain, dripSequence = "iterate", ease = "bounce",
  dripSpeed = 1500, iterationSpeedX = 100, jitterWidth = 0)
}
\arguments{
\item{d3rain}{An object of class d3rain}

\item{ease}{Either 'bounce' or 'linear'}

\item{dripSpeed}{Drip speed}

\item{iterationSpeedX}{Iteration speed multiplier}
}
\description{
Adjust drip behavior
}
